requirements.txt
fastapi[standard]
uvicorn
sqlmodel
pymysql
cryptography
pyjwt
passlib[bcrypt]
fastapi-another-jwt-auth
logging
colorlog
python-multipart
boto3